backend/requirements.txt

flask
web3
python-dotenv
